+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `ndh.utils.set_env`
+
 ## [v6.2.0] - 2023-04-06
 
 - Add `ndh.tests.TestNDHLinks`